+++ CHANGES.rst
@@ -2,13 +2,16 @@
  Change History
 ================
 
-5.5.2 (unreleased)
+5.6.0 (unreleased)
 ==================
 
 - Make repozo's recover mode atomic by recovering the backup in a
   temporary file which is then moved to the expected output file.
+
 - Add a new option to repozo in recover mode which allows to verify
   backups integrity on the fly.
+
+- Drop support for Python 3.4.
 
 5.5.1 (2018-10-25)
 ==================